Replaced timeconv.ParseDuration with duration.ParseDuration, removed pkg duration, added strings.Substring
